+++ Service/Service.Designer.cs
@@ -22,13 +22,15 @@
 
 		#region Component Designer generated code
 
-		/// <summary> 
-		/// Required method for Designer support - do not modify 
+		/// <summary>
+		/// Required method for Designer support - do not modify
 		/// the contents of this method with the code editor.
 		/// </summary>
 		private void InitializeComponent()
 		{
-			components = new System.ComponentModel.Container();
+			// 
+			// Service
+			// 
 			this.ServiceName = "SonnenbergService";
 		}