+++ ServiceManager/Program.cs
@@ -9,23 +9,23 @@ namespace Sonnenberg.ServiceManager
     /// The class responsible for messaging starting and stopping commands
     /// invoked by the WindowsService and executed by the ShellServerManager.
     /// </summary>
-    /// <seealso cref="ServerManager.ServerManager"/>
-    /// <seealso cref="ShellServer.ShellServer"/>
-    public class ServiceManager
+    /// <seealso cref="ServerManager.Program"/>
+    /// <seealso cref="ShellServer.Program"/>
+    public class Program
     {
-        private static readonly ILog log = LogManager.GetLogger(typeof(ServiceManager));
+        private static readonly ILog log = LogManager.GetLogger(typeof(Program));
 
         /// <summary>
         /// Instantiates the ShellServerManager and creates an Instance of the ShellServer
         /// to start the ShellServer via ShellServerManager.
         /// </summary>
-        /// <seealso cref="ServerManager.ServerManager"/>
-        /// <seealso cref="ShellServer.ShellServer"/>
+        /// <seealso cref="ServerManager.Program"/>
+        /// <seealso cref="ShellServer.Program"/>
         public void StartShellServer()
         {
-            var shellServerManager = new ServerManager.ServerManager();
+            var shellServerManager = new ServerManager.Program();
 
-            using (var shellServer = new ShellServer.ShellServer())
+            using (var shellServer = new ShellServer.Program())
             {
                 try
                 {
@@ -45,13 +45,13 @@ namespace Sonnenberg.ServiceManager
         /// Instantiates the ShellServerManager and creates an Instance of the ShellServer
         /// to stop the ShellServer via ShellServerManager.
         /// </summary>
-        /// <seealso cref="ServerManager.ServerManager" />
-        /// <seealso cref="ShellServer.ShellServer"/>
+        /// <seealso cref="ServerManager.Program" />
+        /// <seealso cref="ShellServer.Program"/>
         public void StopShellServer()
         {
-            var shellServerManager = new ServerManager.ServerManager();
+            var shellServerManager = new ServerManager.Program();
 
-            using (var shellServer = new ShellServer.ShellServer())
+            using (var shellServer = new ShellServer.Program())
             {
                 try
                 {