+++ CHANGELOG.md
@@ -1,10 +1,16 @@
+## Version 1.1.0.0
+
+### Changed
+- Added support for shell shortcuts
+- Resolves conflicting class file names
+
 ## Version 1.0.1.0
 
 ### Changed
 - Updated project files to SharpShell version 2.7.2
 
 ### Added
-- Set the Name attribute of each ToolStripMenuItem to one and the same name to prevent multiple display of context menu items.
+- Set the Name attribute of each ToolStripMenuItem to one and the same name to prevent multiple display of context menu items
 
 ## Version 1.0.0.0
 First commit.